requirements.txt
asgiref==3.8.1
sqlparse==0.5.0

beautifulsoup4~=4.12.3
python-dotenv~=1.0.1
chardet~=5.2.0
requests~=2.31.0
